+++ app/Livewire/Members/Import/FieldSync.php
@@ -32,7 +32,7 @@ class FieldSync extends Component
     public function showSyncOverview()
     {
         $this->validate();
-        //remove empty values from associative array
+        // remove empty values from associative array
         $this->fieldMap = array_filter($this->fieldMap, function ($value, $key) {
             return is_numeric($value);
         }, ARRAY_FILTER_USE_BOTH);