+++ deploy.example.php
@@ -6,7 +6,7 @@ require 'recipe/laravel.php';
 
 // Config
 
-set('repository', 'git@gitlab.com:mvkematen/club-scheduler.git'); //todo change to github
+set('repository', 'git@gitlab.com:mvkematen/club-scheduler.git'); // todo change to github
 
 add('shared_files', [
     'public/favicon.ico',
@@ -20,7 +20,7 @@ set('writable_recursive', true);
 // Hosts
 import('deployer-hosts.yaml');
 
-//necessary for some hosting systems
+// necessary for some hosting systems
 desc('Creates the symbolic links configured for the application');
 task('artisan:storage:linkRelative', artisan('storage:link --relative', ['min' => 5.3]));